Move common query sections to query help, add pageInfo stuff
