+++ src/schema/ecommerce/set_order_payment_mutation.ts
@@ -6,7 +6,7 @@ import {
 } from "graphql"
 import { OrderReturnType } from "schema/ecommerce/types/order_return"
 import { mutationWithClientMutationId } from "graphql-relay"
-import { RequestedFulfillmentFragment } from "./types/requested_fulfillment_union_type"
+import { RequestedFulfillmentFragment } from "./query_helpers"
 import gql from "lib/gql"
 
 const SetOrderPaymentInputType = new GraphQLInputObjectType({
